+++ app/src/main/java/com/june0122/sunflower/ui/list/UserListViewModelFactory.kt
@@ -1,8 +1,7 @@
-package com.june0122.sunflower.viewmodel
+package com.june0122.sunflower.ui.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.june0122.sunflower.ui.adapter.UserListAdapter
 
 class UserListViewModelFactory(private val userListAdapter: UserListAdapter) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
