+++ app/src/main/res/layout/fragment_bookmark.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.fragment.BookmarkFragment">
+    tools:context=".ui.bookmark.BookmarkFragment">
 
     <TextView
         android:id="@+id/textView"
