+++ app/src/main/res/layout/fragment_user_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.fragment.UserListFragment">
+    tools:context=".ui.list.UserListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_plant_list"
