+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/user_list_dest"
-        android:name="com.june0122.sunflower.ui.fragment.UserListFragment"
+        android:name="com.june0122.sunflower.ui.list.UserListFragment"
         android:label="User List"
         tools:layout="@layout/fragment_user_list">
         <action
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/detail_dest"
-        android:name="com.june0122.sunflower.ui.fragment.UserDetailFragment"
+        android:name="com.june0122.sunflower.ui.detail.UserDetailFragment"
         android:label="User Detail"
         tools:layout="@layout/fragment_user_detail" >
         <argument
             android:name="userData"
-            app:argType="com.june0122.sunflower.model.data.User" />
+            app:argType="com.june0122.sunflower.data.entity.User" />
         <argument
             android:name="bookmarkStatus"
             app:argType="boolean"
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/bookmark_dest"
-        android:name="com.june0122.sunflower.ui.fragment.BookmarkFragment"
+        android:name="com.june0122.sunflower.ui.bookmark.BookmarkFragment"
         android:label="Bookmark"
         tools:layout="@layout/fragment_bookmark" >
